requirements.txt
# Writing Agent - 依赖包配置（严格版本）
# Python 3.10 (64位)

# ==================== LangChain核心 ====================
langchain==0.2.17
langchain-openai==0.1.22
langchain-redis==0.1.1

# ==================== MCP协议 ====================
mcp==1.20.0
langchain-mcp-adapters==0.1.0

# ==================== 数据验证 ====================
pydantic==2.10.0

# ==================== 基础工具 ====================
python-dotenv==1.0.1
requests==2.32.3

# ==================== 浏览器自动化（MCP搜索）====================
selenium==4.15.2

# ==================== 文本处理 ====================
nltk==3.8.1
jieba==0.42.1

# ==================== Redis存储 ====================
redis==5.0.7

# ==================== Web服务 ====================
uvicorn==0.23.2
fastapi==0.103.0
